pyproject.toml
[project]
name = "img-to-webp"
version = "0.1.0"
description = "A Python package that provides a command-line interface (CLI) for converting images to the WebP format. It supports various image formats and allows for resizing images based on predefined rules."
readme = "README.md"
requires-python = "<4.0,>=3.9"
dependencies = [
    "pyyaml>=6.0.0,<7.0.0",
    "pillow<12.0.0,>=11.0.0",
]

[project.scripts]
img-to-webp = "img_to_webp:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "parameterized>=0.9.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
]
lint = [
    "ruff>=0.11.2",
]
